# WIP: Remove SliceDictionaryStreamReader local buffers and performance improvement