--- a/presto-orc/src/main/java/com/facebook/presto/orc/reader/DecimalStreamReader.java
+++ b/presto-orc/src/main/java/com/facebook/presto/orc/reader/DecimalStreamReader.java
@@ -25,7 +25,10 @@ import com.facebook.presto.orc.stream.LongInputStream;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.type.DecimalType;
+import com.facebook.presto.spi.type.Decimals;
 import com.facebook.presto.spi.type.Type;
+import com.facebook.presto.spi.type.UnscaledDecimal128Arithmetic;
+import io.airlift.slice.Slice;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nonnull;
@@ -34,14 +37,14 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.List;
 
-import static com.facebook.presto.orc.OrcReader.MAX_BATCH_SIZE;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.DATA;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.PRESENT;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.SECONDARY;
 import static com.facebook.presto.orc.stream.MissingInputStreamSource.missingStreamSource;
+import static com.facebook.presto.spi.type.UnscaledDecimal128Arithmetic.rescale;
 import static com.google.common.base.MoreObjects.toStringHelper;
+import static com.google.common.base.Verify.verify;
 import static io.airlift.slice.SizeOf.sizeOf;
-import static java.lang.Math.min;
 import static java.util.Objects.requireNonNull;
 
 public class DecimalStreamReader
@@ -94,74 +97,70 @@ public class DecimalStreamReader
             throws IOException
     {
         DecimalType decimalType = (DecimalType) type;
+        int targetScale = decimalType.getScale();
 
         if (!rowGroupOpen) {
             openRowGroup();
         }
 
         seekToOffset();
-        assureVectorSize();
 
         BlockBuilder builder = decimalType.createBlockBuilder(null, nextBatchSize);
-        while (nextBatchSize > 0) {
-            int subBatchSize = min(nextBatchSize, MAX_BATCH_SIZE);
-            if (presentStream == null) {
-                if (decimalStream == null) {
-                    throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but decimal stream is not present");
-                }
-                if (scaleStream == null) {
-                    throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but scale stream is not present");
-                }
-                scaleStream.nextLongVector(subBatchSize, scaleVector);
+
+        if (presentStream == null) {
+            if (decimalStream == null) {
+                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but decimal stream is not present");
+            }
+            if (scaleStream == null) {
+                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but scale stream is not present");
+            }
+
+            for (int i = 0; i < nextBatchSize; i++) {
+                long sourceScale = scaleStream.next();
                 if (decimalType.isShort()) {
-                    decimalStream.nextShortDecimalVector(subBatchSize, builder, decimalType, scaleVector);
+                    long rescaledDecimal = Decimals.rescale(decimalStream.nextLong(), (int) sourceScale, decimalType.getScale());
+                    decimalType.writeLong(builder, rescaledDecimal);
                 }
                 else {
-                    decimalStream.nextLongDecimalVector(subBatchSize, builder, decimalType, scaleVector);
+                    Slice decimal = UnscaledDecimal128Arithmetic.unscaledDecimal();
+                    Slice rescaledDecimal = UnscaledDecimal128Arithmetic.unscaledDecimal();
+
+                    decimalStream.nextLongDecimal(decimal);
+                    rescale(decimal, (int) (decimalType.getScale() - sourceScale), rescaledDecimal);
+                    decimalType.writeSlice(builder, rescaledDecimal);
                 }
             }
-            else {
-                int nullValues = presentStream.getUnsetBits(subBatchSize, nullVector);
-                if (nullValues != subBatchSize) {
-                    if (decimalStream == null) {
-                        throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but decimal stream is not present");
-                    }
-                    if (scaleStream == null) {
-                        throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but scale stream is not present");
-                    }
-
-                    scaleStream.nextLongVector(subBatchSize, scaleVector, nullVector);
-
+        }
+        else {
+            for (int i = 0; i < nextBatchSize; i++) {
+                if (presentStream.nextBit()) {
+                    // The current row is not null
+                    verify(decimalStream != null);
+                    verify(scaleStream != null);
+                    long sourceScale = scaleStream.next();
                     if (decimalType.isShort()) {
-                        decimalStream.nextShortDecimalVector(subBatchSize, builder, decimalType, scaleVector, nullVector);
+                        long rescaledDecimal = Decimals.rescale(decimalStream.nextLong(), (int) sourceScale, decimalType.getScale());
+                        decimalType.writeLong(builder, rescaledDecimal);
                     }
                     else {
-                        decimalStream.nextLongDecimalVector(subBatchSize, builder, decimalType, scaleVector, nullVector);
+                        Slice decimal = UnscaledDecimal128Arithmetic.unscaledDecimal();
+                        Slice rescaledDecimal = UnscaledDecimal128Arithmetic.unscaledDecimal();
+
+                        decimalStream.nextLongDecimal(decimal);
+                        rescale(decimal, (int) (decimalType.getScale() - sourceScale), rescaledDecimal);
+                        decimalType.writeSlice(builder, rescaledDecimal);
                     }
                 }
                 else {
-                    for (int i = 0; i < subBatchSize; i++) {
-                        builder.appendNull();
-                    }
+                    builder.appendNull();
                 }
             }
-            nextBatchSize -= subBatchSize;
         }
 
         readOffset = 0;
         nextBatchSize = 0;
 
         return builder.build();
-    }
-
-    private void assureVectorSize()
-    {
-        int requiredVectorLength = min(nextBatchSize, MAX_BATCH_SIZE);
-        if (nullVector.length < requiredVectorLength) {
-            nullVector = new boolean[requiredVectorLength];
-            scaleVector = new long[requiredVectorLength];
-            systemMemoryContext.setBytes(getRetainedSizeInBytes());
-        }
     }
 
     private void openRowGroup()