--- a/presto-orc/src/main/java/com/facebook/presto/orc/reader/LongDirectStreamReader.java
+++ b/presto-orc/src/main/java/com/facebook/presto/orc/reader/LongDirectStreamReader.java
@@ -32,13 +32,12 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.List;
 
-import static com.facebook.presto.orc.OrcReader.MAX_BATCH_SIZE;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.DATA;
 import static com.facebook.presto.orc.metadata.Stream.StreamKind.PRESENT;
 import static com.facebook.presto.orc.stream.MissingInputStreamSource.missingStreamSource;
 import static com.google.common.base.MoreObjects.toStringHelper;
+import static com.google.common.base.Verify.verify;
 import static io.airlift.slice.SizeOf.sizeOf;
-import static java.lang.Math.min;
 import static java.util.Objects.requireNonNull;
 
 public class LongDirectStreamReader
@@ -109,23 +108,14 @@ public class LongDirectStreamReader
             dataStream.nextLongVector(type, nextBatchSize, builder);
         }
         else {
-            assureVectorSize();
-
-            while (nextBatchSize > 0) {
-                int subBatchSize = min(nextBatchSize, MAX_BATCH_SIZE);
-                int nullValues = presentStream.getUnsetBits(subBatchSize, nullVector);
-                if (nullValues != subBatchSize) {
-                    if (dataStream == null) {
-                        throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
-                    }
-                    dataStream.nextLongVector(type, subBatchSize, builder, nullVector);
+            for (int i = 0; i < nextBatchSize; i++) {
+                if (presentStream.nextBit()) {
+                    verify(dataStream != null);
+                    type.writeLong(builder, dataStream.next());
                 }
                 else {
-                    for (int i = 0; i < subBatchSize; i++) {
-                        builder.appendNull();
-                    }
+                    builder.appendNull();
                 }
-                nextBatchSize -= subBatchSize;
             }
         }
 
@@ -133,15 +123,6 @@ public class LongDirectStreamReader
         nextBatchSize = 0;
 
         return builder.build();
-    }
-
-    private void assureVectorSize()
-    {
-        int requiredVectorLength = min(nextBatchSize, MAX_BATCH_SIZE);
-        if (nullVector.length < requiredVectorLength) {
-            nullVector = new boolean[requiredVectorLength];
-            systemMemoryContext.setBytes(getRetainedSizeInBytes());
-        }
     }
 
     private void openRowGroup()