We used to use local buffers in the StreamReader's to temporarily hold
data temporarily read from input streams, e.g. the isNullVector and
the inDictionaryVector. The memory used by these local buffers is
not neglegible because there could be many StreamReader objects on a
worker node. These buffers could be eliminated by the following
changes:

1) When the underlying input streams are null, we don't need to keep
the buffers;
2) Reuse each value in these buffers when processing the data across
multiple streams.

It helps to improve the performance because it eliminates multiple
repeating data filtering (branching) on multiple input streams. It
also simplifies the code by removing the subBatchSize logic.

The initial performance testing shows 26% improvements in the
readBlock() calls on a single varchar column with and without nulls:

The readBlock() cost for reading a file with 10000 rows of strings
of length 10 for 1 million times:
		                | no nulls | nulls
----------------------------------------
Before change 	|      42s | 77s
After change 	         |      31s | 58s

JMH benchmark results for reading a file with 1 million rows of
strings of length 10:
Before change   |      0.006 ±  0.001   s/op
After change    |      0.005 ±  0.001   s/op