@dain 2018-08-17
The new shape of `SliceDictionaryStreamReader` looks good.  Reassign the issue back to me when you have this applied to the other places you wanted.